z3str3: move bitvector model construction to theory_str_mc
